# apps/apple/Nouvex/openapi.json
{"openapi":"3.1.0","info":{"title":"Nouvex Mail API","version":"0.0.0"},"components":{"schemas":{},"parameters":{}},"paths":{"/health":{"get":{"operationId":"getHealth","responses":{"200":{"description":"Service is up","content":{"application/json":{"schema":{"type":"object","properties":{"ok":{"type":"boolean"}},"required":["ok"]}}}}}}},"/accounts":{"get":{"operationId":"listAccounts","responses":{"200":{"description":"The user's mail accounts","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"email":{"type":"string"}},"required":["id","email"]}}}}}}},"post":{"operationId":"createAccount","requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"email":{"type":"string","format":"email"},"imapHost":{"type":"string","minLength":1},"imapPort":{"type":"integer","minimum":1,"maximum":65535},"smtpHost":{"type":"string","minLength":1},"smtpPort":{"type":"integer","minimum":1,"maximum":65535},"username":{"type":"string","minLength":1},"password":{"type":"string","minLength":1}},"required":["email","imapHost","imapPort","smtpHost","smtpPort","username","password"]}}}},"responses":{"201":{"description":"Account added, first sync queued","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string"}},"required":["id"]}}}}}}},"/accounts/{accountId}/sync":{"post":{"operationId":"syncAccount","parameters":[{"schema":{"type":"string"},"required":true,"name":"accountId","in":"path"}],"responses":{"202":{"description":"Sync queued"},"404":{"description":"No such account"}}}},"/accounts/{accountId}/changes":{"get":{"operationId":"getChanges","description":"Everything that changed after `since`. Start with 0, then pass the returned version. While `more` is true, ask again right away.","parameters":[{"schema":{"type":"string"},"required":true,"name":"accountId","in":"path"},{"schema":{"type":["integer","null"],"minimum":0},"required":false,"name":"since","in":"query"},{"schema":{"type":"integer","minimum":1,"maximum":1000,"default":500},"required":false,"name":"limit","in":"query"}],"responses":{"200":{"description":"Changes after `since`","content":{"application/json":{"schema":{"type":"object","properties":{"version":{"type":"integer"},"more":{"type":"boolean"},"mailboxes":{"type":"object","properties":{"upserted":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"role":{"type":"string","enum":["inbox","sent","drafts","trash","junk","archive",""]}},"required":["id","name","role"]}},"deleted":{"type":"array","items":{"type":"string"}}},"required":["upserted","deleted"]},"messages":{"type":"object","properties":{"upserted":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string"},"mailboxId":{"type":"string"},"messageId":{"type":"string"},"subject":{"type":"string"},"snippet":{"type":"string"},"fromName":{"type":"string"},"fromAddr":{"type":"string"},"sentAt":{"type":"string","format":"date-time"},"flags":{"type":"array","items":{"type":"string"}},"size":{"type":"integer"}},"required":["id","mailboxId","messageId","subject","snippet","fromName","fromAddr","sentAt","flags","size"]}},"deleted":{"type":"array","items":{"type":"string"}}},"required":["upserted","deleted"]}},"required":["version","more","mailboxes","messages"]}}}},"404":{"description":"No such account"}}}},"/accounts/{accountId}/messages/{messageId}/body":{"get":{"operationId":"getMessageBody","description":"The HTML to show for a message, unsanitized. Render it only in a sandboxed frame after sanitizing.","parameters":[{"schema":{"type":"string"},"required":true,"name":"accountId","in":"path"},{"schema":{"type":"string"},"required":true,"name":"messageId","in":"path"}],"responses":{"200":{"description":"Message body","content":{"application/json":{"schema":{"type":"object","properties":{"html":{"type":"string"}},"required":["html"]}}}},"404":{"description":"No such message"}}}},"/accounts/{accountId}/messages/{messageId}/update":{"post":{"operationId":"updateMessage","description":"Mark a message read or unread, or move it to the trash or the archive. Shows up in /changes once the server has it.","parameters":[{"schema":{"type":"string"},"required":true,"name":"accountId","in":"path"},{"schema":{"type":"string"},"required":true,"name":"messageId","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"seen":{"type":"boolean"},"trash":{"type":"boolean"},"archive":{"type":"boolean"}}}}}},"responses":{"202":{"description":"Change queued"},"404":{"description":"No such message"}}}},"/accounts/{accountId}/send":{"post":{"operationId":"sendMessage","description":"Sends a plain text message. The outcome arrives as an evt.message.sent event on /ws.","parameters":[{"schema":{"type":"string"},"required":true,"name":"accountId","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"to":{"type":"array","items":{"type":"string","format":"email"},"minItems":1},"cc":{"type":"array","items":{"type":"string","format":"email"},"default":[]},"bcc":{"type":"array","items":{"type":"string","format":"email"},"default":[]},"subject":{"type":"string"},"text":{"type":"string"}},"required":["to","subject","text"]}}}},"responses":{"202":{"description":"Queued for sending"},"404":{"description":"No such account"}}}}},"webhooks":{}}